Backend/requirements.txt
fastapi
fastapi-cors
uvicorn
transformers
torch
pillow
datasets
timm
pymysql 
sqlalchemy 
passlib 
bcrypt 
python-jose
